pyproject.toml
[project]
name = "nonebot-plugin-theworld"
version = "0.1.1"
description = "「ザ・ワールド」ッ！ 時よ止まれ！"
authors = [
    {name = "Akirami", email = "Akiramiaya@outlook.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.1", "nonebot-adapter-onebot>=2.2.3"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/A-kirami/nonebot-plugin-theworld"
Repository = "https://github.com/A-kirami/nonebot-plugin-theworld"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
